Feat: CurrentTimeLabel 컴포넌트 리팩토링
