+++ apps/web/app/meetings/_components/Schedule/ScheduleRow/CurrentTimeIndicator.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import { useEffect, useState } from "react";
+import { useCurrentTimePosition } from "@/app/_hooks/useCurrentTimePosition";
+import CurrentTimeLabel from "./CurrentTimeLabel";
 
 interface CurrentTimeIndicatorProps {
   slotWidth: number;
@@ -10,58 +11,23 @@ interface CurrentTimeIndicatorProps {
 
 export default function CurrentTimeIndicator(props: CurrentTimeIndicatorProps): JSX.Element {
   const { slotWidth, startHour, endHour } = props;
-  const [currentPosition, setCurrentPosition] = useState<number | null>(null);
-  const [currentTime, setCurrentTime] = useState<string>("");
-
-  useEffect(() => {
-    const updatePosition = (): void => {
-      const now = new Date();
-      const currentMinutes = now.getHours() * 60 + now.getMinutes();
-      const totalMinutes = (endHour - startHour) * 60;
-
-      if (currentMinutes < startHour * 60 || currentMinutes >= endHour * 60) {
-        setCurrentPosition(null);
-        setCurrentTime("");
-        return;
-      }
-
-      const relativeMinutes = currentMinutes - startHour * 60;
-      const totalSlots = (endHour - startHour) * 2;
-      const scheduleWidth = slotWidth * totalSlots;
-      const position = (relativeMinutes / totalMinutes) * scheduleWidth;
-
-      setCurrentPosition(position);
-
-      const hours = String(now.getHours()).padStart(2, "0");
-      const minutes = String(now.getMinutes()).padStart(2, "0");
-      setCurrentTime(`${hours}:${minutes}`);
-    };
-
-    updatePosition();
-    const interval = setInterval(updatePosition, 60000);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, [slotWidth, startHour, endHour]);
+  const { currentPosition, currentTime } = useCurrentTimePosition({
+    slotWidth,
+    startHour,
+    endHour,
+  });
 
   if (currentPosition === null) {
     return <div />;
   }
 
   return (
-    <div>
+    <div className="relative">
       <div
         className="border-custom-black md:top-30 absolute top-0 z-30 h-full border-l-2 md:ml-36"
-        style={{ left: `${String(currentPosition)}px` }}
+        style={{ left: `${currentPosition}px` }}
       />
-
-      <div
-        className="text-xs-semibold text-custom-black absolute -bottom-24 z-30 -ml-16 rounded bg-none md:-bottom-20 md:block"
-        style={{ left: `${String(currentPosition)}px` }}
-      >
-        {currentTime}
-      </div>
+      <CurrentTimeLabel position={currentPosition} time={currentTime} />
     </div>
   );
 }
